# app/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DreamTeam Crowdfunding Platform API",
    "description": "This is the API for the DreamTeam Crowdfunding Platform. It provides endpoints for managing users, contributors, entrepreneurs, startups, and contributions.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/startups": {
      "post": {
        "tags": [
          "startups"
        ],
        "summary": "Create a new startup",
        "description": "Create a new startup in the system",
        "operationId": "createStartup",
        "requestBody": {
          "description": "Created startup object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Startup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Startup"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "startups"
        ],
        "summary": "List all startups",
        "description": "Returns a list of all startups",
        "operationId": "listStartups",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Startup"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/startups/{startupId}": {
      "get": {
        "tags": [
          "startups"
        ],
        "summary": "Get startup by ID",
        "description": "Returns a single startup",
        "operationId": "getStartupById",
        "parameters": [
          {
            "name": "startupId",
            "in": "path",
            "description": "ID of startup to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Startup"
                }
              }
            }
          },
          "404": {
            "description": "Startup not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "startups"
        ],
        "summary": "Update an existing startup",
        "description": "Update an existing startup by Id",
        "operationId": "updateStartup",
        "parameters": [
          {
            "name": "startupId",
            "in": "path",
            "description": "ID of startup to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated startup object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Startup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Startup"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Startup not found"
          }
        }
      },
      "delete": {
        "tags": [
          "startups"
        ],
        "summary": "Delete a startup",
        "description": "Delete a startup",
        "operationId": "deleteStartup",
        "parameters": [
          {
            "name": "startupId",
            "in": "path",
            "description": "Startup id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Startup not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Startup": {
        "type": "object",
        "required": [
          "title",
          "active_status",
          "start_date",
          "end_date",
          "description",
          "video_link",
          "category_id",
          "region_id"
        ],
        "properties": {
          "startup_id": {
            "type": "integer",
            "format": "int64",
            "description": "The auto-generated id of the startup"
          },
          "title": {
            "type": "string",
            "description": "The title of the startup"
          },
          "active_status": {
            "type": "boolean",
            "description": "Whether the startup is active or not"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "description": "The start date of the startup"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "description": "The end date of the startup"
          },
          "description": {
            "type": "string",
            "description": "The description of the startup"
          },
          "video_link": {
            "type": "object",
            "description": "The video link for the startup"
          },
          "donated_amount": {
            "type": "number",
            "description": "The amount donated to the startup"
          },
          "number_of_contributors": {
            "type": "integer",
            "description": "The number of contributors to the startup"
          },
          "rating": {
            "type": "integer",
            "description": "The rating of the startup"
          },
          "type": {
            "type": "string",
            "enum": [
              "charity",
              "equity"
            ],
            "description": "The type of the startup"
          },
          "batch": {
            "type": "string",
            "enum": [
              "close_to_the_goal",
              "just_launched",
              "finished",
              "none"
            ],
            "description": "The batch status of the startup"
          },
          "category_id": {
            "type": "integer",
            "format": "int64",
            "description": "The id of the category the startup belongs to"
          },
          "region_id": {
            "type": "integer",
            "format": "int64",
            "description": "The id of the region the startup is in"
          }
        }
      },
...
    }
  }
}